+++ src/lib/logger.ts
@@ -1,9 +1,7 @@
-// src/lib/logger.ts
-import { Sentry } from "@/lib/sentry";
+import * as Sentry from "@sentry/nextjs";
 
 type LogMeta = Record<string, unknown>;
 
-// Configuration des logs basée sur les variables d'environnement
 const shouldLog = () => {
   return (
     process.env.NEXT_PUBLIC_ENABLE_LOGS === "true" ||
@@ -11,35 +9,30 @@ const shouldLog = () => {
   );
 };
 
-// Logger intelligent avec Sentry externe
 export const log = {
-  // Logs d'info seulement si activés → Scalingo uniquement
   info: (message: string, meta?: LogMeta) => {
     if (shouldLog()) {
       console.log(`📝 ${message}`, meta);
     }
   },
 
-  // Warnings seulement si activés → Scalingo uniquement
   warn: (message: string, meta?: LogMeta) => {
     if (shouldLog()) {
       console.warn(`⚠️ ${message}`, meta);
     }
   },
 
-  // Erreurs → Scalingo + Sentry
   error: (message: string, meta?: LogMeta) => {
-    console.error(`❌ ${message}`, meta); // → Logs Scalingo
+    console.error(`❌ ${message}`, meta);
 
     if (process.env.NODE_ENV === "production") {
       Sentry.captureMessage(`${message}`, {
         level: "error",
         extra: meta,
-      }); // → Sentry
+      });
     }
   },
 
-  // Actions critiques avec métriques → Scalingo uniquement
   critical: (action: string, meta?: LogMeta) => {
     if (shouldLog()) {
       const metrics = {
@@ -52,31 +45,22 @@ export const log = {
     }
   },
 
-  // Debug seulement en dev → Scalingo uniquement
-  debug: (message: string, meta?: LogMeta) => {
-    if (process.env.NODE_ENV === "development" || shouldLog()) {
-      console.debug(`🐛 ${message}`, meta);
-    }
-  },
-
-  // Logs critiques → Scalingo + Sentry breadcrumb
   always: (message: string, meta?: LogMeta) => {
     const data = {
       timestamp: new Date().toISOString(),
       ...meta,
     };
-    console.log(`🚨 ${message}`, data); // → Logs Scalingo
+    console.log(`🚨 ${message}`, data);
 
     if (process.env.NODE_ENV === "production") {
       Sentry.addBreadcrumb({
         message: message,
         data: meta,
         level: "info",
-      }); // → Sentry breadcrumb
+      });
     }
   },
 
-  // Exception avec stack trace → Scalingo + Sentry
   exception: (error: Error, context?: string) => {
     console.error(`💥 EXCEPTION${context ? ` [${context}]` : ""}:`, error);
 
@@ -89,5 +73,4 @@ export const log = {
   },
 };
 
-// Helper pour vérifier si les logs sont activés
 export const isLoggingEnabled = shouldLog;
